requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
httptools==0.6.4
idna==3.10
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
